feat: support creating a block in the Wasm
This is used initially for containing all of the code in a function, so
that all exits from the function can easily exit the block and execute
the function postlude before leaving the function. Doing this required a
significant refactoring, to allow us to the `InstrSeqBuilder` around and
avoid problems from the borrow checker.