benches: add poc2 contract to suite
This commit adds a benchmark for a contract that is particularly poorly
behaved using the interpreter. The original contract has a *much* longer
list definition, but since not even the interpreter allows for its
execution, we shorten it, varying its length as a parameter.

The WebAssembly runtime seriously outshines the interpreter in the
execution of this contract - to the tune of 25x faster.