cost-tracking: scaffold compiler and runtime code
This commit introduces the scaffold necessary to implement cost-tracking
code generation.

The runtime gets access to three new items: the `AccessCostMeter` and
`CostLinker` traits, and the `CostMeter` struct, which together allow
for setting and retrieving the cost limits when running a module. The
runtime also gets a choice of whether to compile Clarity with or without
the cost tracking code.

The binaries - `clar2wasm` and `crosscheck` - can now receive a new
`--cost-tracking` flag, to optionally generate cost tracking code.

To facilitate further implementation, implementers of `SimpleWord` and
`ComplexWord` are extended with the `WordCharge` and trait. This allows
the implementer to simply call `fn charge(n: Scalar)` on the word, to
generate the appropriate cost-tracking code for that word. The semantics
of the scaling factor `n` is word-dependant, and may be either a constant
32-bit integer or a `LocalId` which resolves into a 32-bit integer.

See-also: #616