Add `Word` trait
The `Word` trait separates `ComplexWord::name` and `SimpleWord::name`
into a separate trait, allowing for functionality to be implemented only
based on a word's name, without having to bring in `SimpleWord::visit`
or `ComplexWord::traverse` with it.

This is useful for any case where the above is required, and since we
already have a case for it in cost-tracking, this is the time to
introduce it.

See-also: #616